Use systemd drop-in instead of editting unit files
For the tpm2-abrmd systemd service, we were using ansible to edit the
file in place. A better approach is to use systemd drop-in files so we
can just override a few attributes in a targetted way that won't be
disrupted if the system updates the unit file.

Fixes issue #39